+++ src/main/java/io/quarkus/gizmo2/impl/InterfaceCreatorImpl.java
@@ -7,6 +7,7 @@ import java.util.function.Consumer;
 
 import io.github.dmlloyd.classfile.ClassBuilder;
 import io.github.dmlloyd.classfile.extras.reflect.AccessFlag;
+import io.quarkus.gizmo2.ClassOutput;
 import io.quarkus.gizmo2.creator.AbstractMethodCreator;
 import io.quarkus.gizmo2.creator.InstanceMethodCreator;
 import io.quarkus.gizmo2.creator.InterfaceCreator;
@@ -15,7 +16,7 @@ import io.quarkus.gizmo2.desc.MethodDesc;
 
 public final class InterfaceCreatorImpl extends TypeCreatorImpl implements InterfaceCreator {
 
-    InterfaceCreatorImpl(final ClassDesc type, final ClassOutputImpl output, final ClassBuilder zb) {
+    InterfaceCreatorImpl(final ClassDesc type, final ClassOutput output, final ClassBuilder zb) {
         super(type, output, zb, AccessFlag.INTERFACE.mask()
                 | AccessFlag.ABSTRACT.mask()
                 | AccessFlag.SYNTHETIC.mask()
